requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pymysql==1.1.1
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
PyJWT==2.10.1
nanoid==2.0.0
loguru==0.7.3
pandas==2.2.3
